+++ pypit/ararc.py
@@ -1,21 +1,25 @@
 from __future__ import (print_function, absolute_import, division, unicode_literals)
 
+import os
+import time
+import inspect
+
 import numpy as np
+from matplotlib import pyplot as plt
+from matplotlib import gridspec, font_manager
+
+import arclines
+#from arclines.holy.grail import basic, semi_brute, general
+
 from pypit import arpca
 from pypit import arparse as settings
-from pypit import armsgs
+from pypit import msgs
 from pypit import arsave
 from pypit import arutils
 from pypit import ararclines
 from pypit import arqa
-from matplotlib import pyplot as plt
-import os
-import time
-
 from pypit import ardebug as debugger
-
-# Logging
-msgs = armsgs.get_logger()
+from pypit import arcyarc
 
 
 def detect_lines(slf, det, msarc, censpec=None, MK_SATMASK=False):
@@ -53,7 +57,6 @@ def detect_lines(slf, det, msarc, censpec=None, MK_SATMASK=False):
       The spectrum used to find detections. This spectrum has
       had any "continuum" emission subtracted off
     """
-    from pypit import arcyarc
     # Extract a rough spectrum of the arc in each order
     msgs.info("Detecting lines")
     msgs.info("Extracting an approximate arc spectrum at the centre of the chip")
@@ -567,7 +570,8 @@ def simple_calib(slf, det, get_poly=False):
         xrej=xrej, yrej=yrej, mask=mask, spec=yprep, nrej=aparm['nsig_rej_final'],
         shift=0., tcent=tcent)
     # QA
-    arqa.arc_fit_qa(slf, final_fit)
+#    arqa.arc_fit_qa(slf, final_fit)
+    arc_fit_qa(slf, final_fit)
     # RMS
     rms_ang = arutils.calc_fit_rms(xfit, yfit, fit, aparm['func'], minv=fmin, maxv=fmax)
     wave = arutils.func_val(fit, np.arange(slf._msarc[det-1].shape[0])/float(slf._msarc[det-1].shape[0]),
@@ -593,7 +597,6 @@ def calib_with_arclines(slf, det, get_poly=False, use_method="general"):
     final_fit : dict
       Dict of fit info
     """
-    from arclines.holy.grail import basic, semi_brute, general
     # Parameters (just for convenience)
     aparm = slf._arcparam[det-1]
     # Extract the arc
@@ -601,13 +604,15 @@ def calib_with_arclines(slf, det, get_poly=False, use_method="general"):
     tampl, tcent, twid, w, satsnd, spec = detect_lines(slf, det, slf._msarc[det-1])
 
     if use_method == "semi-brute":
-        best_dict, final_fit = semi_brute(spec, aparm['lamps'], aparm['wv_cen'], aparm['disp'], fit_parm=aparm, min_ampl=aparm['min_ampl'])
+        best_dict, final_fit = arclines.holy.grail.semi_brute(spec, aparm['lamps'], aparm['wv_cen'], aparm['disp'], fit_parm=aparm, min_ampl=aparm['min_ampl'])
     elif use_method == "basic":
-        stuff = basic(spec, aparm['lamps'], aparm['wv_cen'], aparm['disp'])
+        stuff = arclines.holy.grail.basic(spec, aparm['lamps'], aparm['wv_cen'], aparm['disp'])
         status, ngd_match, match_idx, scores, final_fit = stuff
-    else: # Now preferred
-        best_dict, final_fit = general(spec, aparm['lamps'], fit_parm=aparm, min_ampl=aparm['min_ampl'])
-    arqa.arc_fit_qa(slf, final_fit)
+    else:
+        # Now preferred
+        best_dict, final_fit = arclines.holy.grail.general(spec, aparm['lamps'], fit_parm=aparm, min_ampl=aparm['min_ampl'])
+#    arqa.arc_fit_qa(slf, final_fit)
+    arc_fit_qa(slf, final_fit)
     #
     return final_fit
 
@@ -678,4 +683,114 @@ def new_saturation_mask(a, satlevel):
                 mask = determine_saturation_region(a, x, y, -1, -1, satdown, satlevel, mask)
 
     return mask.astype(int)
+
+
+def arc_fit_qa(slf, fit, outfile=None, ids_only=False, title=None):
+    """
+    QA for Arc spectrum
+
+    Parameters
+    ----------
+    fit : Wavelength fit
+    arc_spec : ndarray
+      Arc spectrum
+    outfile : str, optional
+      Name of output file
+    """
+
+    plt.rcdefaults()
+    plt.rcParams['font.family']= 'times new roman'
+
+    # Grab the named of the method
+    method = inspect.stack()[0][3]
+    # Outfil
+    if outfile is None:
+        outfile = arqa.set_qa_filename(slf.setup, method)
+    #
+    arc_spec = fit['spec']
+
+    # Begin
+    if not ids_only:
+        plt.figure(figsize=(8, 4.0))
+        plt.clf()
+        gs = gridspec.GridSpec(2, 2)
+        idfont = 'xx-small'
+    else:
+        plt.figure(figsize=(11, 8.5))
+        plt.clf()
+        gs = gridspec.GridSpec(1, 1)
+        idfont = 'small'
+
+    # Simple spectrum plot
+    ax_spec = plt.subplot(gs[:,0])
+    ax_spec.plot(np.arange(len(arc_spec)), arc_spec)
+    ymin, ymax = 0., np.max(arc_spec)
+    ysep = ymax*0.03
+    for kk, x in enumerate(fit['xfit']*fit['xnorm']):
+        yline = np.max(arc_spec[int(x)-2:int(x)+2])
+        # Tick mark
+        ax_spec.plot([x,x], [yline+ysep*0.25, yline+ysep], 'g-')
+        # label
+        ax_spec.text(x, yline+ysep*1.3,
+            '{:s} {:g}'.format(fit['ions'][kk], fit['yfit'][kk]), ha='center', va='bottom',
+            size=idfont, rotation=90., color='green')
+    ax_spec.set_xlim(0., len(arc_spec))
+    ax_spec.set_ylim(ymin, ymax*1.2)
+    ax_spec.set_xlabel('Pixel')
+    ax_spec.set_ylabel('Flux')
+    if title is not None:
+        ax_spec.text(0.04, 0.93, title, transform=ax_spec.transAxes,
+                     size='x-large', ha='left')#, bbox={'facecolor':'white'})
+    if ids_only:
+        plt.tight_layout(pad=0.2, h_pad=0.0, w_pad=0.0)
+        plt.savefig(outfile, dpi=800)
+        plt.close()
+        return
+
+    # Arc Fit
+    ax_fit = plt.subplot(gs[0, 1])
+    # Points
+    ax_fit.scatter(fit['xfit']*fit['xnorm'], fit['yfit'], marker='x')
+    if len(fit['xrej']) > 0:
+        ax_fit.scatter(fit['xrej']*fit['xnorm'], fit['yrej'], marker='o',
+            edgecolor='gray', facecolor='none')
+    # Solution
+    xval = np.arange(len(arc_spec))
+    wave = arutils.func_val(fit['fitc'], xval/fit['xnorm'], 'legendre',
+        minv=fit['fmin'], maxv=fit['fmax'])
+    ax_fit.plot(xval, wave, 'r-')
+    xmin, xmax = 0., len(arc_spec)
+    ax_fit.set_xlim(xmin, xmax)
+    ymin,ymax = np.min(wave)*.95,  np.max(wave)*1.05
+    ax_fit.set_ylim(np.min(wave)*.95,  np.max(wave)*1.05)
+    ax_fit.set_ylabel('Wavelength')
+    ax_fit.get_xaxis().set_ticks([]) # Suppress labeling
+    # Stats
+    wave_fit = arutils.func_val(fit['fitc'], fit['xfit'], 'legendre',
+        minv=fit['fmin'], maxv=fit['fmax'])
+    rms = np.sqrt(np.sum((fit['yfit']-wave_fit)**2)/len(fit['xfit'])) # Ang
+    dwv_pix = np.median(np.abs(wave-np.roll(wave,1)))
+    ax_fit.text(0.1*len(arc_spec), 0.90*ymin+(ymax-ymin),
+        r'$\Delta\lambda$={:.3f}$\AA$ (per pix)'.format(dwv_pix), size='small')
+    ax_fit.text(0.1*len(arc_spec), 0.80*ymin+(ymax-ymin),
+        'RMS={:.3f} (pixels)'.format(rms/dwv_pix), size='small')
+    # Arc Residuals
+    ax_res = plt.subplot(gs[1,1])
+    res = fit['yfit']-wave_fit
+    ax_res.scatter(fit['xfit']*fit['xnorm'], res/dwv_pix, marker='x')
+    ax_res.plot([xmin,xmax], [0.,0], 'k--')
+    ax_res.set_xlim(xmin, xmax)
+    ax_res.set_xlabel('Pixel')
+    ax_res.set_ylabel('Residuals (Pix)')
+
+    # Finish
+    plt.tight_layout(pad=0.2, h_pad=0.0, w_pad=0.0)
+    plt.savefig(outfile, dpi=800)
+    plt.close()
+
+    plt.rcdefaults()
+
+    return
+
+