+++ pypit/tests/test_pypit.py
@@ -1,4 +1,8 @@
 # Module to run tests on arsave
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import os
 import pytest